add aggregate / GROUP BY tests

+++ Tests/TestHelpers.cs
@@ -11,7 +11,7 @@ namespace Tests
             Engines.TestTable tt = Engines.TestTableBuilder.NewBuilder()
                 .WithTableName("mytable")
                 .WithColumnNames(new string[] { "keycolumn", "city_name", "state_code", "population" })
-                .WithColumnTypes(new ExpressionOperandType[] { ExpressionOperandType.INTEGER, ExpressionOperandType.VARCHAR, ExpressionOperandType.VARCHAR, ExpressionOperandType.INTEGER })
+                .WithColumnTypes(new ExpressionOperandType[] { ExpressionOperandType.INTEGER, ExpressionOperandType.VARCHAR, ExpressionOperandType.VARCHAR, ExpressionOperandType.DECIMAL })
                 .WithRow(new object[] { 1, "Monroeville", "PA", 25000 })
                 .WithRow(new object[] { 2, "Sammamish", "WA", 37000 })
                 .WithRow(new object[] { 3, "New York", "NY", 11500000 })
